Removed AddCard component
