--- src/components/List/AddCard/AddCard.js
@@ -1,111 +1,0 @@
-import React, { useState, useRef, useEffect } from "react";
-import PropTypes from "prop-types";
-import { FiPlus, FiX } from "react-icons/fi";
-import { useDispatch } from "react-redux";
-import { addCard } from "../../../features/cardsSlice.js";
-
-const AddCard = ({ listId }) => {
-  const dispatch = useDispatch();
-  const [open, setOpen] = useState(false);
-  const [text, setText] = useState("");
-  const inputRef = useRef(null);
-
-  useEffect(() => {
-    if (open) {
-      inputRef.current.focus();
-    }
-  }, [open]);
-
-  const handleOnChange = (e) => {
-    setText(e.target.value);
-  };
-
-  const handleAddCard = () => {
-    if (text.trim.length !== 0) { 
-      setOpen(false);
-      setText("");
-      dispatch(addCard({ listId, content: text }));
-    }
-  };
-
-  const handleKeyDown = (e) => {
-    switch (e.code) {
-      case 'Enter':
-        handleAddCard();
-        break;
-      case 'Escape':
-        handleClose();
-        break;
-    }
-  }
-
-
-  const handleClose = () => {
-    setOpen(false);
-    setText("");
-  };
-
-  return (
-    <>
-      <Collapse isOpen={open}>
-        <textarea
-          ref={inputRef}
-          value={text}
-          placeholder="Type something..."
-          autoFocus
-          className="w-full box-border outline-none border-2 border-trello-blue-100 shadow bg-white hover:bg-trello-gray-400 rounded-md p-2 my-1.5"
-          // onBlur={() => setOpen(false)}
-          onChange={handleOnChange}
-          onKeyDown={handleKeyDown}
-        />
-        <div className="flex items-center">
-          <button
-            type="button"
-            className="py-2 px-3 bg-trello-green-100 hover:bg-trello-green-200 text-white transition ease-in duration-200 text-center text-base shadow-md rounded-md"
-            onClick={handleAddCard}
-          >
-            Add Card
-          </button>
-          <FiX
-            onClick={handleClose}
-            size={36}
-            className="text-trello-gray-200 hover:bg-trello-gray-500 cursor-pointer rounded-full ml-1 p-1"
-          />
-        </div>
-      </Collapse>
-
-      <Collapse isOpen={!open}>
-        <div
-          className="flex hover:bg-trello-gray-500 rounded-ibsm items-center cursor-pointer text-trello-gray-600 p-1"
-          onClick={() => setOpen(true)}
-        >
-          <FiPlus size={20} />
-          <h2 className="flex-1 ml-1.5">Add a card</h2>
-        </div>
-      </Collapse>
-    </>
-  );
-};
-
-AddCard.propTypes = {
-  listId: PropTypes.string,
-};
-
-export default AddCard;
-
-const Collapse = ({ isOpen, children }) => {
-  const ref = useRef(null);
-
-  const inlineStyle = isOpen ? { height: ref.current?.scrollHeight } : { height: 0 };
-
-  return (
-    <div ref={ref} aria-hidden={!isOpen} style={inlineStyle} className="transition-height ease overflow-hidden duration-300">
-      {children}
-    </div>
-  );
-};
-
-Collapse.propTypes = {
-  isOpen: PropTypes.bool,
-  children: PropTypes.oneOfType([PropTypes.arrayOf(PropTypes.node), PropTypes.node]),
-};
